Make GUI trainer acessible via an entry point
