+++ setup.py
@@ -31,4 +31,9 @@ setup(
     url="https://github.com/sdatkinson/",
     install_requires=requirements,
     packages=find_packages(),
+    entry_points={
+        'console_scripts': [
+            'nam = nam.train.gui:run',
+        ]
+    }
 )
